+++ README.md
@@ -1,8 +1,8 @@
 # excavator ♠️
 
-excavator is a `spadingScript` for KoLmafia (at least r20145) for collaboratively gathering data while playing the Kingdom of Loathing. It collects sends data via kmail to the player `Excavator` when possible (i.e. when you're not in an encounter), otherwise it is cached in the `spadingData` property. The cache can be processed by running the `spade` command in the CLI.
+excavator is a `spadingScript` for KoLmafia (at least r20145) for collaboratively gathering data while playing the Kingdom of Loathing.
 
-Only data pertinent to the [current projects](https://excavator.loathers.net/projects) are reported (outcomes from combats, state of flags or counters on your player etc) and each packet of data can be reviewed and approved before it is sent. No personal information will ever be transmitted.
+Only data pertinent to the [current projects](https://excavator.loathers.net/projects) are reported (outcomes from combats, state of flags or counters on your player etc) and each packet of data can be reviewed and approved before it is sent. No personal information beyond your player id will ever be transmitted.
 
 These data are pulled every 15 minutes to a database and displayed on https://excavator.loathers.net. Before this site they were imported to [Google Sheets](https://tinyurl.com/excavator-data), where some data remains for posterity.
