+++ package.json
@@ -13,8 +13,7 @@
   "scripts": {
     "build": "yarn workspace excavator-script run build",
     "format": "yarn workspaces foreach -Ap run format && yarn prettier --write .github",
-    "lint": "yarn workspaces foreach -Ap run lint && yarn prettier --check .github",
-    "etl": "yarn workspace excavator-web run etl"
+    "lint": "yarn workspaces foreach -Ap run lint && yarn prettier --check .github"
   },
   "engines": {
     "node": ">= 18.20"
